move definitions out of client/index into their own files
