+++ src/client/markdown.ts
@@ -5,7 +5,8 @@
  * repository.
  */
 
-import { IPageData, IPageNode } from "./index";
+import { IBlock } from "./compiler";
+import { INavigable } from "./tags";
 
 export interface IMarkdownPluginData {
     /**
@@ -19,4 +20,33 @@ export interface IMarkdownPluginData {
     pages: {
         [reference: string]: IPageData;
     };
+}
+
+/**
+ * A single Documentalist page, parsed from a single source file.
+ */
+export interface IPageData extends IBlock {
+    /** Unique identifier for addressing this page. */
+    reference: string;
+
+    /** Fully qualified route to this page: slash-separated references of all parent pages. */
+    route: string;
+
+    /** Human-friendly title of this page. */
+    title: string;
+}
+
+/** An `@#+` tag belongs to a specific page. */
+export interface IHeadingNode extends INavigable {
+    /** Display title of page heading. */
+    title: string;
+}
+
+/** A page has ordered children composed of `@#+` and `@page` tags. */
+export interface IPageNode extends IHeadingNode {
+    /** Ordered list of pages and headings that appear on this page. */
+    children: Array<IPageNode | IHeadingNode>;
+
+    /** Unique reference of this page, used for retrieval from store. */
+    reference: string;
 }
