+++ src/client/plugin.ts
@@ -5,42 +5,7 @@
  * repository.
  */
 
-import { IBlock } from "../client";
-
-/**
- * Abstract representation of a file, containing absolute path and synchronous `read` operation.
- */
-export interface IFile {
-    path: string;
-    read(): string;
-}
-
-/**
- * Each plugin receives a `Compiler` instance to aid in the processing of source files.
- */
-export interface ICompiler {
-    /**
-     * Converts an array of entries into a map of key to entry, using given
-     * callback to extract key from each item.
-     */
-    objectify<T>(array: T[], getKey: (item: T) => string): { [key: string]: T };
-
-    /**
-     * Render a block of content by extracting metadata (YAML front matter) and
-     * splitting text content into markdown-rendered HTML strings and `{ tag,
-     * value }` objects.
-     *
-     * To prevent special strings like "@include" from being parsed, a reserved
-     * tag words array may be provided, in which case the line will be left as
-     * is.
-     */
-    renderBlock(blockContent: string, reservedTagWords?: string[]): IBlock;
-
-    /**
-     * Render a string of markdown to HTML, using the options from `Documentalist`.
-     */
-    renderMarkdown(markdown: string): string;
-}
+import { ICompiler } from "./compiler";
 
 /**
  * A Documentalist plugin is an object with a `compile(files, compiler)` function
@@ -71,4 +36,13 @@ export interface ICompiler {
  */
 export interface IPlugin<T> {
     compile(files: IFile[], compiler: ICompiler): T | Promise<T>;
+}
+
+/**
+ * Abstract representation of a file, containing absolute path and synchronous `read` operation.
+ * This allows plugins to use only the path of a file without reading it.
+ */
+export interface IFile {
+    path: string;
+    read(): string;
 }
